backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.14.0
pymysql==1.1.1
cryptography==44.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-dotenv==1.0.1

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Media
Pillow==11.0.0
aiofiles==24.1.0

# Security
bcrypt==4.2.1

# Utilities
python-dateutil==2.9.0
python-slugify==8.0.4

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
